S3/S3A1-Gemini/requirements.txt
Flask
requests
PyPDF2
pytube
python-dotenv